# src/app/api/tasks/route.ts
import { NextRequest, NextResponse } from "next/server";
import { prisma } from "@/lib/prisma";
import { verifyToken } from "@/lib/auth";
import { CreateTaskSchema } from "@/lib/schemas/task";
import { sendTelegramMessage } from "@/lib/telegram";

/**
 * Получить все задачи пользователя
 * @description Возвращает список задач текущего пользователя
 * @auth bearer
 * @response 200:TaskArrayResponse:Список задач
 * @response 401:ErrorResponse:Не авторизован
 * @openapi
 */

export async function GET(request: NextRequest) {
  try {
    const authHeader = request.headers.get("authorization");
    const token = authHeader?.split(" ")[1];

    if (!token) {
      return NextResponse.json({ error: "Не авторизован" }, { status: 401 });
    }

    const payload = verifyToken(token);
    if (!payload) {
      return NextResponse.json({ error: "Неверный токен" }, { status: 401 });
    }

    const tasks = await prisma.task.findMany({
      where: { userId: payload.userId },
      include: { project: true },
      orderBy: { createdAt: "desc" },
    });

    return NextResponse.json(tasks);
  } catch (error) {
    console.error("❌ GET /api/tasks error:", error);
    return NextResponse.json(
      { error: "Ошибка загрузки задач" },
      { status: 500 },
    );
  }
}

/**
 * Создать задачу
 * @description Создаёт новую задачу с валидацией через Zod
 * @auth bearer
 * @body CreateTaskSchema
 * @response 201:TaskResponse:Задача создана
 * @response 400:ErrorResponse:Неверные данные
 * @response 401:ErrorResponse:Не авторизован
 * @openapi
 */

export async function POST(request: NextRequest) {
  try {
    const authHeader = request.headers.get("authorization");
    const token = authHeader?.split(" ")[1];

    if (!token) {
      return NextResponse.json({ error: "Не авторизован" }, { status: 401 });
    }

    const payload = verifyToken(token);
    if (!payload) {
      return NextResponse.json({ error: "Неверный токен" }, { status: 401 });
    }

    const body = await request.json();
    const result = CreateTaskSchema.safeParse(body);

    if (!result.success) {
      return NextResponse.json(
        { error: "Неверные данные", details: result.error.issues },
        { status: 400 },
      );
    }

    const { title, description, status, priority, deadline, projectId } =
      result.data;

    const now = new Date();
    const taskData: any = {
      title,
      description,
      status,
      priority,
      deadline: deadline ? new Date(deadline) : null,
      projectId: projectId || null,
      userId: payload.userId,
    };

    if (status === "IN_PROGRESS") {
      taskData.startedAt = now;
    }
    if (status === "DONE") {
      taskData.startedAt = now;
      taskData.completedAt = now;
    }

    const task = await prisma.task.create({
      data: taskData,
      include: { project: true },
    });

    if (deadline) {
      const user = await prisma.user.findUnique({
        where: { id: payload.userId },
        select: { telegramChatId: true },
      });

      if (user?.telegramChatId) {
        try {
          await sendTelegramMessage(
            user.telegramChatId,
            `📌 Новая задача с дедлайном:\n<b>${title}</b>\nДедлайн: ${new Date(deadline).toLocaleString("ru-RU")}`,
          );
        } catch (err) {
          console.error("Ошибка отправки в Telegram:", err);
        }
      }
    }

    return NextResponse.json(task, { status: 201 });
  } catch (error) {
    console.error("❌ POST /api/tasks error:", error);
    return NextResponse.json(
      { error: "Ошибка создания задачи" },
      { status: 500 },
    );
  }
}